rewrote HTML pipeline to use built in renderMarkup command
seems to work much better than my original attempts at using @joplin/htmlRenderer

unlike when I use @joplin/renderer on its own, the renderMarkup command adheres to all of the joplin syntax settings on the markdown tab.

This eliminates a huge amount of complexity with managing different markdown-it plugins to match joplin's settings.

The one downside is that the markdown-it github alerts support no longer works (because joplin doesn't use that setting), though I might be able to do something with a markdown-it content script (it's not exactly a critical feature in any case).